test serial parser. remove unused module. rename confirmation module as it is already nested
