+++ web/src/components/Upload/util.test.tsx
@@ -3,7 +3,8 @@ import * as fs from 'fs'
 import {
   getDustrakStart,
   getDustrakEnd,
-  getGpsStart
+  getGpsStart,
+  getDustrakSerial
 } from 'components/Upload/util'
 
 const testDir: string = `${__dirname}/test-data`
@@ -52,6 +53,14 @@ describe('parses gps and dustrak files', () => {
       missingStartDustrak,
       empty
     ] = await Promise.all(filePromises)
+  })
+
+  it('should find the dustrak serial', () => {
+    expect(getDustrakSerial(validDustrak)).toEqual('8530094612')
+  })
+
+  it('should handle a corrupted dustrak file with blank serial', () => {
+    expect(getDustrakSerial(empty)).toEqual('')
   })
 
   it('should find the time listed in the GPS', () => {
